pre-interview-2021-22/day30/searchinrotated/program_test.go
package program

import "testing"

func Test_search(t *testing.T) {
	type args struct {
		nums   []int
		target int
	}
	tests := []struct {
		name string
		args args
		want int
	}{
		// // TODO: Add test cases.
		{
			name: "test 1",
			args: args{
				nums:   []int{4, 5, 6, 7, 0, 1, 2},
				target: 0,
			},
			want: 4,
		},
		{
			name: "test 2",
			args: args{
				nums:   []int{7, 8, 0, 1, 3, 4, 5},
				target: 0,
			},
			want: 2,
		},
		{
			name: "test 3",
			args: args{
				nums:   []int{3, 1},
				target: 1,
			},
			want: 1,
		},
	}
	for _, tt := range tests {
		t.Run(tt.name, func(t *testing.T) {
			if got := search(tt.args.nums, tt.args.target); got != tt.want {
				t.Errorf("search() = %v, want %v", got, tt.want)
			}
		})
	}
}
